Added absract command to fix inheritance chain
